Rename Branch, terminate work on ch. 19 & 20 for now
Rename branch to reflect work on challenge 19 and 20
done here in one.
There is a many time pad tool already:
https://github.com/CameronLonsdale/MTP/
So it's a bit of a waste of time to write it from scratch.
Do at some point in time: write a proper statistical analysis
for cipher text (Chars and Bigrams), needed for challenge 20.
MTP only looks for spaces. So maybe provide the extended analysis
in a pull request there too?
Also now working in PyCharm. Added .idea files. Do on master as well.